Extract ServiceLoader class from ServicesManagerServiceClass
